Refactor handlers.py to utilize ExpenseHandler and AnalyticsHandler classes for improved organization and modularity. Removed redundant user data handling and streamlined message handler registration.
